#RI-5462 - Dedicated editor for sql
