+++ configs/webpack.config.web.common.ts
@@ -71,7 +71,7 @@ export default {
     new HtmlWebpackPlugin({ template: 'index.html.ejs' }),
 
     new MonacoWebpackPlugin({
-      languages: ['yaml', 'typescript', 'javascript', 'json'],
+      languages: ['yaml', 'typescript', 'javascript', 'json', 'sql'],
       customLanguages: [
         {
           label: 'yaml',
