+++ redisinsight/ui/src/components/monaco-editor/components/dedicated-editor/DedicatedEditor.spec.tsx
