+++ redisinsight/ui/src/components/monaco-editor/components/dedicated-editor/index.ts
@@ -1,0 +1,3 @@
+import DedicatedEditor from './DedicatedEditor'
+
+export default DedicatedEditor
