+++ redisinsight/ui/src/components/monaco-editor/components/dedicated-editor/styles.module.scss
