+++ redisinsight/ui/src/components/monaco-editor/index.ts
@@ -2,10 +2,12 @@ import MonacoEditor from './MonacoEditor'
 import MonacoJS from './components/monaco-js'
 import MonacoJson from './components/monaco-json'
 import MonacoYaml from './components/monaco-yaml'
+import DedicatedEditor from './components/dedicated-editor'
 
 export {
   MonacoEditor,
   MonacoJS,
   MonacoJson,
   MonacoYaml,
+  DedicatedEditor,
 }
