+++ redisinsight/ui/src/constants/monaco/monaco.ts
@@ -1,22 +1,46 @@
 import { monaco as monacoEditor } from 'react-monaco-editor'
+import { getCypherCompletionProvider } from 'uiSrc/utils/monaco/cypher/completionProvider'
+import { getCypherMonarchTokensProvider } from 'uiSrc/utils/monaco/cypher/monarchTokensProvider'
+import { cypherLanguageConfiguration } from './cypher'
+import { DSL, DSLNaming } from '../commands'
 
 export interface MonacoSyntaxLang {
   [key: string]: {
     name: string
     id: string
-    config: monacoEditor.languages.LanguageConfiguration,
-    completionProvider: () => monacoEditor.languages.CompletionItemProvider,
-    tokensProvider: () => monacoEditor.languages.IMonarchLanguage
+    config?: monacoEditor.languages.LanguageConfiguration,
+    completionProvider?: () => monacoEditor.languages.CompletionItemProvider,
+    tokensProvider?: () => monacoEditor.languages.IMonarchLanguage
   }
 }
 
 export enum MonacoLanguage {
   Redis = 'redisLanguage',
-  Cypher = 'cypherLanguage'
+  Cypher = 'cypherLanguage',
+  SQL = 'sql',
+  Text = 'text',
 }
 
 export const DEFAULT_MONACO_YAML_URI = 'http://example.com/schema-name.json'
 export const DEFAULT_MONACO_FILE_MATCH = '*'
+
+export const DEDICATED_EDITOR_LANGUAGES: MonacoSyntaxLang = {
+  [DSL.cypher]: {
+    name: DSLNaming[DSL.cypher],
+    id: MonacoLanguage.Cypher,
+    config: cypherLanguageConfiguration,
+    completionProvider: getCypherCompletionProvider,
+    tokensProvider: getCypherMonarchTokensProvider,
+  },
+  [DSL.sql]: {
+    name: DSLNaming[DSL.sql],
+    id: MonacoLanguage.SQL,
+  },
+  [DSL.jmespath]: {
+    name: DSLNaming[DSL.jmespath],
+    id: MonacoLanguage.Text,
+  },
+}
 
 export const MONACO_MANUAL = '// Workbench is the advanced Redis command-line interface that allows to send commands to Redis, read and visualize the replies sent by the server.\n'
   + '// Enter multiple commands at different rows to run them at once.\n'
